+++ src/main/java/net/tysontheember/emberstextapi/EmbersTextAPI.java
@@ -28,6 +28,7 @@ import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
 import org.slf4j.Logger;
 import net.minecraftforge.network.PacketDistributor;
+import net.tysontheember.emberstextapi.config.ClientConfig;
 import net.tysontheember.emberstextapi.immersivemessages.api.ImmersiveMessage;
 import net.tysontheember.emberstextapi.immersivemessages.network.TooltipPacket;
 import net.tysontheember.emberstextapi.network.Network;
@@ -52,6 +53,8 @@ public class EmbersTextAPI
 
         // Register the item to a creative tab
         modEventBus.addListener(this::addCreative);
+
+        context.registerConfig(ModConfig.Type.CLIENT, ClientConfig.SPEC);
     }
 
     private void commonSetup(final FMLCommonSetupEvent event)